+++ src/main/java/com/dementor/domain/mentor/service/MentorService.java
@@ -11,7 +11,6 @@ import com.dementor.domain.member.repository.MemberRepository;
 import com.dementor.domain.mentor.dto.request.MentorApplyProposalRequest;
 import com.dementor.domain.mentor.dto.request.MentorApplyStatusRequest;
 import com.dementor.domain.mentor.dto.request.MentorChangeRequest;
-import com.dementor.domain.mentor.dto.request.MentorUpdateRequest;
 import com.dementor.domain.mentor.dto.response.MentorApplyResponse;
 import com.dementor.domain.mentor.dto.response.MentorApplyStatusResponse;
 import com.dementor.domain.mentor.dto.response.MentorChangeResponse;
@@ -24,13 +23,14 @@ import com.dementor.domain.mentor.repository.MentorRepository;
 import com.dementor.domain.mentorapplyproposal.entity.MentorApplyProposal;
 import com.dementor.domain.mentorapplyproposal.entity.MentorApplyProposalStatus;
 import com.dementor.domain.mentorapplyproposal.repository.MentorApplyProposalRepository;
+import com.dementor.domain.mentoreditproposal.dto.MentorEditProposalRequest;
+import com.dementor.domain.mentoreditproposal.dto.MentorUpdateRequest;
 import com.dementor.domain.mentoreditproposal.entity.MentorEditProposal;
 import com.dementor.domain.mentoreditproposal.entity.MentorEditProposalStatus;
 import com.dementor.domain.mentoreditproposal.repository.MentorEditProposalRepository;
 import com.dementor.domain.postattachment.repository.PostAttachmentRepository;
 import com.dementor.domain.postattachment.service.PostAttachmentService;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -49,17 +49,17 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class MentorService {
-    private final MentorRepository mentorRepository;
-    private final MemberRepository memberRepository;
-    private final JobRepository jobRepository;
-    private final PostAttachmentRepository attachmentRepository;
-    private final MentorEditProposalRepository mentorEditProposalRepository;
-    private final MentorApplyProposalRepository mentorApplyProposalRepository;
-    private final ObjectMapper objectMapper;
-    private final PostAttachmentService postAttachmentService;
-    private final ApplyRepository applyRepository;
+	private final MentorRepository mentorRepository;
+	private final MemberRepository memberRepository;
+	private final JobRepository jobRepository;
+	private final PostAttachmentRepository attachmentRepository;
+	private final MentorEditProposalRepository mentorEditProposalRepository;
+	private final MentorApplyProposalRepository mentorApplyProposalRepository;
+	private final ObjectMapper objectMapper;
+	private final PostAttachmentService postAttachmentService;
+	private final ApplyRepository applyRepository;
 
-    //멘토 지원하기
+	//멘토 지원하기
     @Transactional
     public MentorApplyProposal applyMentor(MentorApplyProposalRequest.MentorApplyProposalRequestDto requestDto) {
         // 회원 엔티티 조회
@@ -67,38 +67,38 @@ public class MentorService {
                 .orElseThrow(() -> new MentorException(MentorErrorCode.MENTOR_NOT_FOUND,
                         "회원을 찾을 수 없습니다: " + requestDto.memberId()));
 
-        // 회원의 역할이 이미 MENTOR인지 확인
-        if (member.getUserRole() == UserRole.MENTOR) {
-            throw new MentorException(MentorErrorCode.MENTOR_ALREADY_EXISTS,
-                    "이미 멘토로 등록된 사용자입니다: " + requestDto.memberId());
-        }
+		// 회원의 역할이 이미 MENTOR인지 확인
+		if (member.getUserRole() == UserRole.MENTOR) {
+			throw new MentorException(MentorErrorCode.MENTOR_ALREADY_EXISTS,
+				"이미 멘토로 등록된 사용자입니다: " + requestDto.memberId());
+		}
 
-        // 직무 엔티티 조회
-        Job job = jobRepository.findById(requestDto.jobId())
-                .orElseThrow(() -> new MentorException(MentorErrorCode.INVALID_MENTOR_APPLICATION,
-                        "직무를 찾을 수 없습니다: " + requestDto.jobId()));
+		// 직무 엔티티 조회
+		Job job = jobRepository.findById(requestDto.jobId())
+			.orElseThrow(() -> new MentorException(MentorErrorCode.INVALID_MENTOR_APPLICATION,
+				"직무를 찾을 수 없습니다: " + requestDto.jobId()));
 
-        // 이미 지원 내역이 있는지 확인
-        if (mentorApplyProposalRepository.existsByMemberId(requestDto.memberId())) {
-            throw new MentorException(MentorErrorCode.INVALID_MENTOR_APPLICATION,
-                    "이미 멘토 지원 내역이 존재합니다: " + requestDto.memberId());
-        }
+		// 이미 지원 내역이 있는지 확인
+		if (mentorApplyProposalRepository.existsByMemberId(requestDto.memberId())) {
+			throw new MentorException(MentorErrorCode.INVALID_MENTOR_APPLICATION,
+				"이미 멘토 지원 내역이 존재합니다: " + requestDto.memberId());
+		}
 
-        // 멘토 애플리케이션 엔티티 생성 - 초기 상태는 PENDING
-        MentorApplyProposal mentorApplyProposal = MentorApplyProposal.builder()
-                .member(member)
-                .job(job)
-                .name(requestDto.name())
-                .career(requestDto.career())
-                .phone(requestDto.phone())
-                .email(requestDto.email())
-                .currentCompany(requestDto.currentCompany())
-                .introduction(requestDto.introduction())
-                .status(MentorApplyProposalStatus.PENDING)
-                .build();
+		// 멘토 애플리케이션 엔티티 생성 - 초기 상태는 PENDING
+		MentorApplyProposal mentorApplyProposal = MentorApplyProposal.builder()
+			.member(member)
+			.job(job)
+			.name(requestDto.name())
+			.career(requestDto.career())
+			.phone(requestDto.phone())
+			.email(requestDto.email())
+			.currentCompany(requestDto.currentCompany())
+			.introduction(requestDto.introduction())
+			.status(MentorApplyProposalStatus.PENDING)
+			.build();
 
-        // 멘토 애플리케이션 저장 (ID 생성)
-        MentorApplyProposal savedApplication = mentorApplyProposalRepository.save(mentorApplyProposal);
+		// 멘토 애플리케이션 저장 (ID 생성)
+		MentorApplyProposal savedApplication = mentorApplyProposalRepository.save(mentorApplyProposal);
 
         // 첨부파일 연결
         if (requestDto.attachmentId() != null && !requestDto.attachmentId().isEmpty()) {
@@ -131,42 +131,47 @@ public class MentorService {
                     "이미 정보 수정 요청 중입니다: " + memberId);
         }
 
-        // 변경 사항 추출
-        Map<String, Map<String, Object>> changes = extractChanges(mentor, requestDto);
+		// 변경 사항이 있는지 확인
+		if (!requestDto.hasChanges(mentor)) {
+			throw new MentorException(MentorErrorCode.INVALID_MENTOR_APPLICATION,
+				"변경된 내용이 없습니다.");
+		}
 
-        // 변경 사항을 JSON으로 변환
-        String changesJson;
-        try {
-            changesJson = objectMapper.writeValueAsString(changes);
-        } catch (JsonProcessingException e) {
-            throw new MentorException(MentorErrorCode.INVALID_MENTOR_APPLICATION,
-                    "변경 내용을 JSON으로 변환하는데 실패했습니다: " + e.getMessage());
-        }
+		// 변경 사항 추출
+		Map<String, Map<String, Object>> changes = extractChanges(mentor, requestDto);
 
-        // 수정 요청 엔티티 생성 및 저장
-        MentorEditProposal modification = MentorEditProposal.builder()
-                .member(mentor.getMember())
-                .changes(changesJson)
-                .status(MentorEditProposalStatus.PENDING)
-                .build();
+		// 변경 사항을 JSON으로 변환
+		String changesJson;
+		try {
+			changesJson = objectMapper.writeValueAsString(changes);
+		} catch (JsonProcessingException e) {
+			throw new MentorException(MentorErrorCode.INVALID_MENTOR_APPLICATION,
+				"변경 내용을 JSON으로 변환하는데 실패했습니다: " + e.getMessage());
+		}
 
-        MentorEditProposal savedModification = mentorEditProposalRepository.save(modification);
+		// 수정 요청 엔티티 생성 및 저장
+		MentorEditProposal modification = MentorEditProposal.builder()
+			.member(mentor.getMember())
+			.status(MentorEditProposalStatus.PENDING)
+			.build();
 
-        // 첨부 파일 처리
-        if (requestDto.attachmentId() != null && !requestDto.attachmentId().isEmpty()) {
-            for (Long attachmentId : requestDto.attachmentId()) {
-                // 첨부 파일 존재 여부 확인
-                attachmentRepository.findById(attachmentId)
-                        .ifPresent(attachment -> {
-                            if (!attachment.getMember().getId().equals(mentor.getMember().getId())) {
-                                throw new MentorException(MentorErrorCode.UNAUTHORIZED_ACCESS,
-                                        "본인이 업로드한 파일만 연결할 수 있습니다: " + attachmentId);
-                            }
-                            attachment.connectToMentorModification(savedModification);
-                            attachmentRepository.save(attachment);
-                        });
-            }
-        }
+		MentorEditProposal savedModification = mentorEditProposalRepository.save(modification);
+
+		// 첨부 파일 처리
+		if (requestDto.attachmentId() != null && !requestDto.attachmentId().isEmpty()) {
+			for (Long attachmentId : requestDto.attachmentId()) {
+				// 첨부 파일 존재 여부 확인
+				attachmentRepository.findById(attachmentId)
+					.ifPresent(attachment -> {
+						if (!attachment.getMember().getId().equals(mentor.getMember().getId())) {
+							throw new MentorException(MentorErrorCode.UNAUTHORIZED_ACCESS,
+								"본인이 업로드한 파일만 연결할 수 있습니다: " + attachmentId);
+						}
+						attachment.connectToMentorModification(savedModification);
+						attachmentRepository.save(attachment);
+					});
+			}
+		}
 
         // 멘토의 수정 상태 업데이트
         mentor.updateModificationStatus(ModificationStatus.PENDING);
@@ -193,161 +198,159 @@ public class MentorService {
         // 완료된 멘토링 수 계산 - 멘토링 신청 날짜가 오늘보다 이전이면 완료된 상태
         Integer completedSessions = applyRepository.countCompletedSessions(classIds);
 
+		return MentorInfoResponse.from(mentor, totalClasses, pendingRequests, completedSessions);
+	}
 
-        return MentorInfoResponse.from(mentor, totalClasses, pendingRequests, completedSessions);
-    }
+	//멘토 정보 수정 요청 목록 조회
+	public MentorChangeResponse.ChangeListResponse getModificationRequests(
+		Long memberId,
+		MentorChangeRequest.ModificationRequestParams params) {
+		// 멘토 존재 여부 확인
+		if (!mentorRepository.existsById(memberId)) {
+			throw new MentorException(MentorErrorCode.MENTOR_NOT_FOUND,
+				"해당 멘토를 찾을 수 없습니다: " + memberId);
+		}
 
-    //멘토 정보 수정 요청 목록 조회
-    public MentorChangeResponse.ChangeListResponse getModificationRequests(
-            Long memberId,
-            MentorChangeRequest.ModificationRequestParams params) {
-        // 멘토 존재 여부 확인
-        if (!mentorRepository.existsById(memberId)) {
-            throw new MentorException(MentorErrorCode.MENTOR_NOT_FOUND,
-                    "해당 멘토를 찾을 수 없습니다: " + memberId);
-        }
+		// 페이지네이션 설정
+		Pageable pageable = PageRequest.of(
+			params.page() - 1, // 0-based page index
+			params.size(),
+			Sort.by(Sort.Direction.DESC, "createdAt")
+		);
 
-        // 페이지네이션 설정
-        Pageable pageable = PageRequest.of(
-                params.page() - 1, // 0-based page index
-                params.size(),
-                Sort.by(Sort.Direction.DESC, "createdAt")
-        );
+		try {
+			// 상태 필터가 있으면 상태별로 조회, 없으면 전체 조회
+			Page<MentorEditProposal> modificationPage;
 
-        try {
-            // 상태 필터가 있으면 상태별로 조회, 없으면 전체 조회
-            Page<MentorEditProposal> modificationPage;
+			if (params.status() != null) {
+				MentorEditProposalStatus status = MentorEditProposalStatus.valueOf(params.status());
+				modificationPage = mentorEditProposalRepository.findByMemberIdAndStatus(memberId, status, pageable);
+			} else {
+				modificationPage = mentorEditProposalRepository.findByMemberId(memberId, pageable);
+			}
 
-            if (params.status() != null) {
-                MentorEditProposalStatus status = MentorEditProposalStatus.valueOf(params.status());
-                modificationPage = mentorEditProposalRepository.findByMemberIdAndStatus(memberId, status, pageable);
-            } else {
-                modificationPage = mentorEditProposalRepository.findByMemberId(memberId, pageable);
-            }
+			// 결과 변환 및 반환
+			List<MentorChangeResponse.ChangeRequestData> changeRequests = modificationPage.getContent().stream()
+				.map(this::convertToChangeRequestData)
+				.collect(Collectors.toList());
 
-            // 결과 변환 및 반환
-            List<MentorChangeResponse.ChangeRequestData> changeRequests = modificationPage.getContent().stream()
-                    .map(this::convertToChangeRequestData)
-                    .collect(Collectors.toList());
+			return new MentorChangeResponse.ChangeListResponse(
+				changeRequests,
+				new MentorChangeResponse.Pagination(
+					params.page(),
+					params.size(),
+					modificationPage.getTotalElements()
+				)
+			);
+		} catch (Exception e) {
+			throw new MentorException(MentorErrorCode.INVALID_STATUS_PARAM,
+				"멘토 정보 수정 요청 목록 조회 중 오류가 발생했습니다: " + e.getMessage());
+		}
+	}
 
-            return new MentorChangeResponse.ChangeListResponse(
-                    changeRequests,
-                    new MentorChangeResponse.Pagination(
-                            params.page(),
-                            params.size(),
-                            modificationPage.getTotalElements()
-                    )
-            );
-        } catch (Exception e) {
-            throw new MentorException(MentorErrorCode.INVALID_STATUS_PARAM,
-                    "멘토 정보 수정 요청 목록 조회 중 오류가 발생했습니다: " + e.getMessage());
-        }
-    }
+	//멘토 정보 수정 요청을 DTO로 변환합니다.
+	private MentorChangeResponse.ChangeRequestData convertToChangeRequestData(MentorEditProposal modification) {
+		Map<String, MentorChangeResponse.FieldChange<?>> modifiedFields = new HashMap<>();
 
-    //멘토 정보 수정 요청을 DTO로 변환합니다.
-    private MentorChangeResponse.ChangeRequestData convertToChangeRequestData(MentorEditProposal modification) {
-        Map<String, MentorChangeResponse.FieldChange<?>> modifiedFields = new HashMap<>();
+		try {
+			// Map<String, Map<String, Object>> changes = objectMapper.readValue(
+			// 	modification.getChanges(),
+			// 	new TypeReference<Map<String, Map<String, Object>>>() {
+			// 	}
+			// );
 
-        try {
-            Map<String, Map<String, Object>> changes = objectMapper.readValue(
-                    modification.getChanges(),
-                    new TypeReference<Map<String, Map<String, Object>>>() {}
-            );
+			// for (Map.Entry<String, Map<String, Object>> entry : changes.entrySet()) {
+			// 	String fieldName = entry.getKey();
+			// 	Map<String, Object> change = entry.getValue();
+			//
+			// 	Object before = change.get("before");
+			// 	Object after = change.get("after");
+			//
+			// 	modifiedFields.put(fieldName, new MentorChangeResponse.FieldChange<>(before, after));
+			// }
+		} catch (Exception e) {
+			// JSON 파싱 실패 시 빈 맵 반환
+		}
 
-            for (Map.Entry<String, Map<String, Object>> entry : changes.entrySet()) {
-                String fieldName = entry.getKey();
-                Map<String, Object> change = entry.getValue();
+		return new MentorChangeResponse.ChangeRequestData(
+			modification.getId(),
+			modification.getStatus().name(),
+			modification.getCreatedAt(),
+			modifiedFields
+		);
+	}
 
-                Object before = change.get("before");
-                Object after = change.get("after");
+	//멘토 엔티티와 수정 요청 DTO를 비교하여 변경 사항을 추출합니다.
+	private Map<String, Map<String, Object>> extractChanges(Mentor mentor,
+		MentorUpdateRequest.MentorUpdateRequestDto dto) {
+		Map<String, Map<String, Object>> changes = new HashMap<>();
 
-                modifiedFields.put(fieldName, new MentorChangeResponse.FieldChange<>(before, after));
-            }
-        } catch (Exception e) {
-            // JSON 파싱 실패 시 빈 맵 반환
-        }
+		// career 변경 확인
+		if (dto.career() != null && !dto.career().equals(mentor.getCareer())) {
+			Map<String, Object> fieldChange = new HashMap<>();
+			fieldChange.put("before", mentor.getCareer());
+			fieldChange.put("after", dto.career());
+			changes.put("career", fieldChange);
+		}
 
-        return new MentorChangeResponse.ChangeRequestData(
-                modification.getId(),
-                modification.getStatus().name(),
-                modification.getCreatedAt(),
-                modifiedFields
-        );
-    }
+		// phone 변경 확인
+		if (dto.phone() != null && !dto.phone().equals(mentor.getPhone())) {
+			Map<String, Object> fieldChange = new HashMap<>();
+			fieldChange.put("before", mentor.getPhone());
+			fieldChange.put("after", dto.phone());
+			changes.put("phone", fieldChange);
+		}
 
-    //멘토 엔티티와 수정 요청 DTO를 비교하여 변경 사항을 추출합니다.
-    private Map<String, Map<String, Object>> extractChanges(Mentor mentor, MentorUpdateRequest.MentorUpdateRequestDto dto) {
-        Map<String, Map<String, Object>> changes = new HashMap<>();
+		// currentCompany 변경 확인
+		if (dto.currentCompany() != null && !dto.currentCompany().equals(mentor.getCurrentCompany())) {
+			Map<String, Object> fieldChange = new HashMap<>();
+			fieldChange.put("before", mentor.getCurrentCompany());
+			fieldChange.put("after", dto.currentCompany());
+			changes.put("currentCompany", fieldChange);
+		}
 
-        // career 변경 확인
-        if (dto.career() != null && !dto.career().equals(mentor.getCareer())) {
-            Map<String, Object> fieldChange = new HashMap<>();
-            fieldChange.put("before", mentor.getCareer());
-            fieldChange.put("after", dto.career());
-            changes.put("career", fieldChange);
-        }
+		// jobId 변경 확인
+		if (dto.jobId() != null && !dto.jobId().equals(mentor.getJob().getId())) {
+			Map<String, Object> fieldChange = new HashMap<>();
+			fieldChange.put("before", mentor.getJob().getId());
+			fieldChange.put("after", dto.jobId());
+			changes.put("jobId", fieldChange);
+		}
 
-        // phone 변경 확인
-        if (dto.phone() != null && !dto.phone().equals(mentor.getPhone())) {
-            Map<String, Object> fieldChange = new HashMap<>();
-            fieldChange.put("before", mentor.getPhone());
-            fieldChange.put("after", dto.phone());
-            changes.put("phone", fieldChange);
-        }
+		// email 변경 확인 - 이제 멘토 엔티티에서 직접 확인
+		if (dto.email() != null && !dto.email().equals(mentor.getEmail())) {
+			Map<String, Object> fieldChange = new HashMap<>();
+			fieldChange.put("before", mentor.getEmail());
+			fieldChange.put("after", dto.email());
+			changes.put("email", fieldChange);
+		}
 
-        // currentCompany 변경 확인
-        if (dto.currentCompany() != null && !dto.currentCompany().equals(mentor.getCurrentCompany())) {
-            Map<String, Object> fieldChange = new HashMap<>();
-            fieldChange.put("before", mentor.getCurrentCompany());
-            fieldChange.put("after", dto.currentCompany());
-            changes.put("currentCompany", fieldChange);
-        }
+		// introduction 변경 확인
+		if (dto.introduction() != null && !dto.introduction().equals(mentor.getIntroduction())) {
+			Map<String, Object> fieldChange = new HashMap<>();
+			fieldChange.put("before", mentor.getIntroduction());
+			fieldChange.put("after", dto.introduction());
+			changes.put("introduction", fieldChange);
+		}
 
-        // jobId 변경 확인
-        if (dto.jobId() != null && !dto.jobId().equals(mentor.getJob().getId())) {
-            Map<String, Object> fieldChange = new HashMap<>();
-            fieldChange.put("before", mentor.getJob().getId());
-            fieldChange.put("after", dto.jobId());
-            changes.put("jobId", fieldChange);
-        }
+		return changes;
+	}
 
-        // email 변경 확인 - 이제 멘토 엔티티에서 직접 확인
-        if (dto.email() != null && !dto.email().equals(mentor.getEmail())) {
-            Map<String, Object> fieldChange = new HashMap<>();
-            fieldChange.put("before", mentor.getEmail());
-            fieldChange.put("after", dto.email());
-            changes.put("email", fieldChange);
-        }
+	@Transactional
+	public MentorApplyResponse.GetApplyMenteePageList getApplyByMentor(Long memberId, int page, int size) {
 
-        // introduction 변경 확인
-        if (dto.introduction() != null && !dto.introduction().equals(mentor.getIntroduction())) {
-            Map<String, Object> fieldChange = new HashMap<>();
-            fieldChange.put("before", mentor.getIntroduction());
-            fieldChange.put("after", dto.introduction());
-            changes.put("introduction", fieldChange);
-        }
+		Mentor mentor = mentorRepository.findById(memberId)
+			.orElseThrow(() -> new IllegalArgumentException("멘토만 조회할 수 있습니다."));
 
+		// 멘토가 가진 클래스 아이디 목록 조회
+		List<Long> classId = mentorRepository.findMentoringClassIdsByMentor(mentor);
 
+		// 멘토가 가진 클래스 아이디 목록으로 신청 목록 조회
+		Page<Apply> applyPage = applyRepository.findByMentoringClassIdIn(classId, PageRequest.of(page, size));
 
-        return changes;
-    }
+		return MentorApplyResponse.GetApplyMenteePageList.from(applyPage, page, size);
 
-    @Transactional
-    public MentorApplyResponse.GetApplyMenteePageList getApplyByMentor(Long memberId, int page, int size) {
-
-        Mentor mentor = mentorRepository.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("멘토만 조회할 수 있습니다."));
-
-
-        // 멘토가 가진 클래스 아이디 목록 조회
-        List<Long> classId = mentorRepository.findMentoringClassIdsByMentor(mentor);
-
-        // 멘토가 가진 클래스 아이디 목록으로 신청 목록 조회
-        Page<Apply> applyPage = applyRepository.findByMentoringClassIdIn(classId, PageRequest.of(page, size));
-
-        return MentorApplyResponse.GetApplyMenteePageList.from(applyPage, page, size);
-
-    }
+	}
 
     @Transactional
     public MentorApplyStatusResponse updateApplyStatus(Long memberId, Long applyId, MentorApplyStatusRequest request) {
