+++ src/main/java/com/dementor/domain/mentoreditproposal/entity/MentorEditProposal.java
@@ -1,9 +1,25 @@
 package com.dementor.domain.mentoreditproposal.entity;
 
+import com.dementor.domain.job.entity.Job;
 import com.dementor.domain.member.entity.Member;
 import com.dementor.global.base.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "mentor_edit_proposal")
@@ -21,8 +37,18 @@ public class MentorEditProposal extends BaseEntity {
     @JoinColumn(name = "member_id")
     private Member member;
 
-    @Column(name = "changes", nullable = false, columnDefinition = "TEXT")
-    private String changes; // JSON 형태로 변경 내용 저장
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "job_id", nullable = false)
+    private Job job;
+
+    @Column(nullable = false)
+    private Integer career;
+
+    @Column(length = 20)
+    private String currentCompany;
+
+    @Column(nullable = false)
+    private String introduction;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
